+++ example/hashlib_example.dart
@@ -1,5 +1,4 @@
 import 'package:hashlib/hashlib.dart';
-import 'package:hashlib/src/codecs_base.dart';
 
 void main() {
   var text = "Happy Hashing!";