+++ lib/src/xxh3.dart
@@ -30,6 +30,12 @@ class XXH3 extends BlockHashBase {
   final int seed;
   final List<int>? secret;
 
+  /// Creates a new instance of [XXH3].
+  ///
+  /// Paramters:
+  /// - [seed] is an optional 64-bit integer. Default: 0
+  /// - [secret] is an array of bytes. The length should be at least 136.
+  /// - If the [secret] is present, the [seed] is ignored.
   const XXH3({this.seed = 0, this.secret});
 
   @override