+++ lib/src/xxh32.dart
@@ -23,6 +23,10 @@ const XXHash32 xxh32 = XXHash32(0);
 class XXHash32 extends BlockHashBase {
   final int seed;
 
+  /// Creates a new instance of [XXHash32].
+  ///
+  /// Paramters:
+  /// - [seed] is an optional 32-bit integer. Default: 0
   const XXHash32([this.seed = 0]);
 
   @override