+++ lib/src/xxh64.dart
@@ -23,6 +23,10 @@ const XXHash64 xxh64 = XXHash64(0);
 class XXHash64 extends BlockHashBase {
   final int seed;
 
+  /// Creates a new instance of [XXHash64].
+  ///
+  /// Paramters:
+  /// - [seed] is an optional 64-bit integer. Default: 0
   const XXHash64([this.seed = 0]);
 
   @override