+++ test/base32_test.dart
@@ -1,7 +1,7 @@
 // Copyright (c) 2023, Sudipto Chandra
 // All rights reserved. Check LICENSE file for details.
 
-import 'package:hashlib/src/codecs_base.dart';
+import 'package:hashlib/hashlib.dart';
 import 'package:test/test.dart';
 
 void main() {